requirements.txt
anthropic>=0.86.0
python-dotenv>=1.2.2
requests>=2.33.0
beautifulsoup4>=4.12.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
twilio>=9.0.0
